Data Pipeline and Processing
Needs revision, but first, we have to look at the database records and structure them appropriately. Also, the number of records per sample need to be 5 and over for correct outlier rejection. Normalisation needs to be applied before ML and Labels should be structured appropriately on a commonly agreed consensus.